Updated video endpoint
