+++ src/templates/index.html
@@ -15,7 +15,7 @@
     </div>
 
     <div class="video-container">
-        <img id="videoStream" src="{{ url_for('routes.video_feed') }}" alt="Live Feed">
+        <img id="videoStream" src="{{ url_for('video_feed') }}" alt="Live Feed">
     </div>
 
     <div class="controller-container">
